Separate context declaration and provider
