+++ frontend/src/context/EmailContext.tsx
@@ -1,25 +1,9 @@
-import { createContext, useContext, useMemo, useState, useEffect } from "react";
+import { useContext, useMemo, useState, useEffect } from "react";
 import { getValidationCode } from "../services/email-api";
 import type { UserId } from "../types/types";
+import { EmailContext } from "./emailContext";
 
-interface EmailContextProps {
-    sendVerificationCode: (email: string) => Promise<boolean>;
-    loading: boolean;
-    setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-    error: string | null;
-    setError: React.Dispatch<React.SetStateAction<string | null>>;
-    setErrorReset: (message: string) => void;
-    email: string | undefined;
-    setEmail: React.Dispatch<React.SetStateAction<string | undefined>>;
-    codeSent: boolean;
-    setCodeSent: React.Dispatch<React.SetStateAction<boolean>>;
-    userId: UserId | undefined;
-    setUserId: React.Dispatch<React.SetStateAction<UserId | undefined>>;
-    isSubscribed: boolean;
-    setIsSubscribed: React.Dispatch<React.SetStateAction<boolean>>;
-};
 
-export const EmailContext = createContext<EmailContextProps | undefined>(undefined);
 
 export const EmailProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
