+++ frontend/src/context/emailContext.ts
@@ -1,0 +1,21 @@
+import { createContext } from "react";
+import type { UserId } from "../types/types";
+
+interface EmailContextProps {
+    sendVerificationCode: (email: string) => Promise<boolean>;
+    loading: boolean;
+    setLoading: React.Dispatch<React.SetStateAction<boolean>>;
+    error: string | null;
+    setError: React.Dispatch<React.SetStateAction<string | null>>;
+    setErrorReset: (message: string) => void;
+    email: string | undefined;
+    setEmail: React.Dispatch<React.SetStateAction<string | undefined>>;
+    codeSent: boolean;
+    setCodeSent: React.Dispatch<React.SetStateAction<boolean>>;
+    userId: UserId | undefined;
+    setUserId: React.Dispatch<React.SetStateAction<UserId | undefined>>;
+    isSubscribed: boolean;
+    setIsSubscribed: React.Dispatch<React.SetStateAction<boolean>>;
+};
+
+export const EmailContext = createContext<EmailContextProps | undefined>(undefined);
